Added Stripping of H1 tags on clean.
This should fix some strange behavior on the Whiskey Media sites. Mainly,
the H1 headers are not supported by parchment (at least you can not select
them from the header drop down), but if you paste text with the header
format (for example copy/pasting the product header from an Amazon page)
the H1 tag will be copied with it into parchment.

This wouldn't be a problem, but for some reason the H1 tag gets completely
removed from the post on postback (server side sanitizing?). This results
in some strange behavoir. What appears to be a perfectly fine post to the
poster, suddenly is missing the pasted text when posted! This is even
stranger on certain pages, because the styling makes it appear as though
the H1 text is regular text! So the user is completely unawares of the
header text.

Example Amazon page to test: http://tinyurl.com/85d3w2t (Try
copying/pasting the product name into parchment and posting it to the
forums)

This is particularly annoying if your trying to post product suggestions
to users of a Tech site :) (I only became aware of this strange behavior
because Rallier [Tested Mod] complained about it in IRC)

Anyways, this fix just strips H1 tags from the content of the parchment
text box on clean. It does this by just moving everything from the inside
of the tag to its parent tag right before where the H1 tag would have
been, and then removes the H1 tag.