Firefox is the bane of my existance.
This change added a few bug fixes to blockquote, but I couldnt fix them
all.

The main bug being fixed was a problem on Chrome, where if you had some
text selected, and wanted to make a link but accidentally hit the quote
block button, and then hit the button again to undo the quote block, the
whole line would be selected instead of the original selected text. Making
it slightly annoying to reselect it.

While fixing that, I realized that the quoteblock plugin is broken for
Firefox. It has two main problems: the first is that trying to do a
quoteblock on the first line on of a post without any other lines didn't
work. This is due to a bug in firefox (the relevant bugzilla link is in
the comments), and is fixed by a simple kludge that adds a second line on
formatting (removing it afterwards).

The second bug is a little more insidious. The problem is that it's
impossible to undo a blockquote in Firefox. This is due to the way Firefox
implements the formatBlock command. Unlike webkit, which replaces any
block element with the formatted block (allowing you to toggle the
format), firefox does that for everything *but* the blockquote object. So
instead of toggling the block quote, the formatBlock command just inserts
an extra paragraph inside the block quote...

The whole BlockQuote plugin also doesn't work at all for Internet
Explorer, since that platform does not support Blockquote as a value ofr
formatBlock. I was unable to fix the Firefox bug in such a way that would
keep the selection bug I first mentioned also fixed. If I have some more
time to work on it, the only method I think will fix this is implementing
a custom version of formatBlock for cross-compatability. This custom
version may also fix the not working on IE thing.

Anyways, firefox is the bane of my existance.